chore: updates workflow to enable npm trusted publishing (#451)
NPM package publishing will be migrated to use Trusted Publishing (via
OIDC).

To ensure security and restrict access, we will configure the OIDC token
request to only be available to jobs running in the dedicated 'release'
environment. This environment will be a required condition for obtaining
the necessary NPM token.

Changes:
* ~~Set node version to 24, which includes npm v11~~ Separate PR.
* Adds `id-token` permission required for creating JWT
* Adds environment `release` for obtaining the short-lived tokens
* Pins external action `ncipollo/release-action` to commit SHA

Co-authored-by: Luca Bertelli <luca.berte98@hotmail.it>